// types/movieTypes.ts
export type Movie = {
    id: number;
    poster_path: string;
    backdrop_path: string;
    title: string;
    popularity: number;
    vote_average: number;
    release_date: string;
    name: string;
    original_name: string;
    overview: string;
    number_of_episodes: number;
    number_of_seasons: number;
    genres: any;
    origin_country: string;
    original_language: string;
    first_air_date: string
    production_countries: any
  };